refactor(data): move PeriodType/UsageSource/UsageFilters to usage-filters.ts, remove threshold
Consolidates shared filter types into usage-filters.ts (alongside the
coercers that already live there) so that usage-controls.tsx can be
deleted. UsageFilters no longer carries a threshold field — threshold is
now URL-only state coerced on read, not part of the shareable filter
shape passed between components.
